+++ src/nvim/api/private/helpers.h
@@ -7,6 +7,7 @@
 #include "nvim/vim.h"
 #include "nvim/getchar.h"
 #include "nvim/memory.h"
+#include "nvim/decoration.h"
 #include "nvim/ex_eval.h"
 #include "nvim/lib/kvec.h"
