+++ src/nvim/buffer_defs.h
@@ -568,7 +568,7 @@ struct file_buffer {
                                 // negative when lines were deleted
   wininfo_T   *b_wininfo;       // list of last used info for each window
   int b_mod_tick_syn;           // last display tick syntax was updated
-  int b_mod_tick_deco;          // last display tick decoration providers
+  int b_mod_tick_decor;         // last display tick decoration providers
                                 // where invoked
 
   long b_mtime;                 // last change time of original file
